Creation of Crops with predefined values.
These values are subject to change as the program becomes more developed.
More likely that modifiers and upgrades can be added. It's just that right now these are being used.
This is the first push and commit for all crops.
As of right now 0 graphical implementation has been made too